refactor: moving select and form select
