+++ apps/bilan-carbone/src/components/emissionFactor/Form/EmissionFactorForm.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import { FormAutocomplete } from '@/components/form/Autocomplete'
-import { FormSelect } from '@/components/form/Select'
 import GlossaryModal from '@/components/modals/GlossaryModal'
 import QualitySelectGroup from '@/components/study/QualitySelectGroup'
 import { customRich } from '@/i18n/customRich'
@@ -11,6 +10,7 @@ import { BCUnit, useUnitLabel } from '@/services/unit'
 import { ManualEmissionFactorUnitList } from '@/utils/emissionFactors'
 import LinkButton from '@abc-transitionbascarbone/components/src/base/LinkButton'
 import LoadingButton from '@abc-transitionbascarbone/components/src/base/LoadingButton'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { FormTextField } from '@abc-transitionbascarbone/components/src/form/TextField'
 import { FormControlLabel, FormLabel, MenuItem, Switch } from '@mui/material'
 import classNames from 'classnames'
