+++ apps/bilan-carbone/src/components/emissionFactor/Form/SubPostSelector.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import { Select } from '@/components/base/Select'
 import { ALL_SUB_POSTS_VALUE } from '@/constants/post.constants'
 import { environmentPostMapping, Post, subPostsByPost } from '@/services/posts'
 import { getSortedPosts } from '@/utils/post'
+import { Select } from '@abc-transitionbascarbone/components/src/base/Select'
 import { Environment, SubPost } from '@abc-transitionbascarbone/db-common/enums'
 import { Checkbox, ListItemText, ListSubheader, MenuItem, SelectChangeEvent } from '@mui/material'
 import { useTranslations } from 'next-intl'
