+++ apps/bilan-carbone/src/components/onboarding/OnboardingStep2.tsx
@@ -1,5 +1,6 @@
 import { OnboardingCommand } from '@/services/serverFunctions/user.command'
 import { getEnvironmentRoles } from '@/utils/user'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { FormTextField } from '@abc-transitionbascarbone/components/src/form/TextField'
 import { Environment, Role } from '@abc-transitionbascarbone/db-common/enums'
 import { Button } from '@abc-transitionbascarbone/ui'
@@ -9,7 +10,6 @@ import { MenuItem } from '@mui/material'
 import classNames from 'classnames'
 import { useTranslations } from 'next-intl'
 import { UseFormReturn, useWatch } from 'react-hook-form'
-import { FormSelect } from '../form/Select'
 import styles from './Onboarding.module.css'
 
 interface Props {
