+++ apps/bilan-carbone/src/components/study/EmissionSourceForm.tsx
@@ -25,6 +25,7 @@ import { formatEmissionFactorNumber, formatNumber } from '@/utils/number'
 import { formatEmissionFromNumber, hasDeprecationPeriod, hasEditionRights, isCAS } from '@/utils/study'
 import { HelpIcon } from '@abc-transitionbascarbone/components'
 import LinkButton from '@abc-transitionbascarbone/components/src/base/LinkButton'
+import { Select } from '@abc-transitionbascarbone/components/src/base/Select'
 import Modal from '@abc-transitionbascarbone/components/src/modals/Modal'
 import type { StudyTag } from '@abc-transitionbascarbone/db-common'
 import {
@@ -50,7 +51,6 @@ import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { useEffect, useMemo, useState } from 'react'
 import { Path } from 'react-hook-form'
-import { Select } from '../base/Select'
 import TagChip from '../base/TagChip'
 import BaseChip from '../emissionFactor/BaseChip'
 import { ImportVersionForFilters } from '../emissionFactor/EmissionFactorsFilters'
