+++ apps/bilan-carbone/src/components/study/new/ExportCheckbox.tsx
@@ -1,8 +1,8 @@
-import { Select } from '@/components/base/Select'
 import GlossaryIconModal from '@/components/modals/GlossaryIconModal'
 import ControlModeChangeWarningModal from '@/components/study/perimeter/ControlModeChangeWarningModal'
 import type { FullStudy } from '@/db/study'
 import { updateStudySpecificExportFields } from '@/services/serverFunctions/study'
+import { Select } from '@abc-transitionbascarbone/components/src/base/Select'
 import { useServerFunction } from '@abc-transitionbascarbone/components/src/hooks/useServerFunction'
 import { ControlMode, Export } from '@abc-transitionbascarbone/db-common/enums'
 import { Checkbox, FormControl, FormControlLabel, MenuItem } from '@mui/material'
