+++ apps/bilan-carbone/src/components/study/organization/Select.tsx
@@ -1,5 +1,4 @@
 'use client'
-import { FormSelect } from '@/components/form/Select'
 import SiteDeselectionWarningModal from '@/components/modals/SiteDeselectionWarningModal'
 import { OrganizationWithSites } from '@/db/account'
 import DynamicSites from '@/environments/core/organization/DynamicSites'
@@ -8,6 +7,7 @@ import { updateOrganizationCommand } from '@/services/serverFunctions/organizati
 import { CreateStudyCommand } from '@/services/serverFunctions/study.command'
 import { CA_UNIT_VALUES, displayCA } from '@/utils/number'
 import Block from '@abc-transitionbascarbone/components/src/base/Block'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { useServerFunction } from '@abc-transitionbascarbone/components/src/hooks/useServerFunction'
 import { Environment, SiteCAUnit } from '@abc-transitionbascarbone/db-common/enums'
 import { NOT_AUTHORIZED } from '@abc-transitionbascarbone/services/permissions/check'
