+++ apps/bilan-carbone/src/components/study/rights/NewStudyRightForm.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import { FormAutocomplete } from '@/components/form/Autocomplete'
-import { FormSelect } from '@/components/form/Select'
 import { getOrganizationVersionAccounts } from '@/db/organization'
 import type { FullStudy } from '@/db/study'
 import { ALREADY_IN_STUDY } from '@/services/permissions/check'
@@ -11,6 +10,7 @@ import { hasSufficientLevel } from '@/utils/study'
 import { isAdmin } from '@/utils/user'
 import Form from '@abc-transitionbascarbone/components/src/base/Form'
 import LoadingButton from '@abc-transitionbascarbone/components/src/base/LoadingButton'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { useServerFunction } from '@abc-transitionbascarbone/components/src/hooks/useServerFunction'
 import { StudyRole } from '@abc-transitionbascarbone/db-common/enums'
 import { useToast } from '@abc-transitionbascarbone/ui'
