+++ apps/bilan-carbone/src/components/study/rights/StudyLevel.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { FormSelect } from '@/components/form/Select'
 import GlossaryModal from '@/components/modals/GlossaryModal'
 import type { FullStudy } from '@/db/study'
 import { customRich } from '@/i18n/customRich'
@@ -8,6 +7,7 @@ import { changeStudyLevel } from '@/services/serverFunctions/study'
 import { ChangeStudyLevelCommand, ChangeStudyLevelCommandValidation } from '@/services/serverFunctions/study.command'
 import { getAllowedLevels } from '@/utils/study'
 import { HelpIcon } from '@abc-transitionbascarbone/components'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { useServerFunction } from '@abc-transitionbascarbone/components/src/hooks/useServerFunction'
 import { Level } from '@abc-transitionbascarbone/db-common/enums'
 import { zodResolver } from '@hookform/resolvers/zod'
