+++ apps/bilan-carbone/src/components/study/transitionPlan/Actions/ActionModalStep2.tsx
@@ -1,8 +1,8 @@
-import { FormSelect } from '@/components/form/Select'
 import GlossaryIconModal from '@/components/modals/GlossaryIconModal'
 import { customRich } from '@/i18n/customRich'
 import { AddActionFormCommand } from '@/services/serverFunctions/action.command'
 import { getOrderedActionRelevances } from '@/utils/action'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { ActionCategory, ActionNature } from '@abc-transitionbascarbone/db-common/enums'
 import { Locale } from '@abc-transitionbascarbone/i18n/config'
 import { Translations } from '@abc-transitionbascarbone/lib'
