+++ apps/bilan-carbone/src/components/study/transitionPlan/LinkingStudyModal.tsx
@@ -1,4 +1,3 @@
-import { Select } from '@/components/base/Select'
 import { FormDatePicker } from '@/components/form/DatePicker'
 import type { FullStudy } from '@/db/study'
 import { getStudyPreviousOccurrences } from '@/services/serverFunctions/study'
@@ -11,6 +10,7 @@ import type { PastStudy } from '@/types/trajectory.types'
 import { convertValue } from '@/utils/study'
 import Form from '@abc-transitionbascarbone/components/src/base/Form'
 import LoadingButton from '@abc-transitionbascarbone/components/src/base/LoadingButton'
+import { Select } from '@abc-transitionbascarbone/components/src/base/Select'
 import { FormTextField } from '@abc-transitionbascarbone/components/src/form/TextField'
 import { useServerFunction } from '@abc-transitionbascarbone/components/src/hooks/useServerFunction'
 import Modal from '@abc-transitionbascarbone/components/src/modals/Modal'
