+++ apps/bilan-carbone/src/environments/base/study/new/Form.tsx
@@ -3,7 +3,6 @@
 import { FormAutocomplete } from '@/components/form/Autocomplete'
 import { FormDatePicker } from '@/components/form/DatePicker'
 import { FormRadio } from '@/components/form/Radio'
-import { FormSelect } from '@/components/form/Select'
 import GlobalNewStudyForm from '@/components/study/new/Form'
 import StudyExportsForm from '@/components/study/perimeter/StudyExportsForm'
 import { getOrganizationVersionAccounts } from '@/db/organization'
@@ -13,6 +12,7 @@ import { getAllowedLevels, hasSufficientLevel } from '@/utils/study'
 import { HelpIcon } from '@abc-transitionbascarbone/components'
 import Block from '@abc-transitionbascarbone/components/src/base/Block'
 import IconLabel from '@abc-transitionbascarbone/components/src/base/IconLabel'
+import { FormSelect } from '@abc-transitionbascarbone/components/src/form/Select'
 import { Level } from '@abc-transitionbascarbone/db-common/enums'
 import { FormControlLabel, MenuItem, Radio } from '@mui/material'
 import { UserSession } from 'next-auth'
