+++ apps/mip/src/app/(dashboard)/equipe/ajouter/page.tsx
@@ -1,7 +1,8 @@
 import withAuth, { UserSessionProps } from '@/components/hoc/withAuth'
+import NewMemberPage from '@/components/pages/NewMember'
 
 const NewMember = ({ user }: UserSessionProps) => {
-  return <p>New Member </p>
+  return <NewMemberPage user={user} />
 }
 
 export default withAuth(NewMember)
